requirements.txt
python-telegram-bot
uuid